IMPALA-966: Attribute type error to the right expression in an insert statement
Currently, if an insert statement contains multiple expressions
that are incompatible with the column type, the error message
returned attributes the error to the wrong expression.
This patch makes sure the right expression is blamed. If there are
multiple incompatible type values for the target column, then the
error is attributed to the first widest (highest precision)
incompatible type expression.

Testing:
- Added tests to AnalyzeStmtsTest.java

Change-Id: I88718fc2cbe1a492165435a542fd2d91d8693a39
Reviewed-on: http://gerrit.cloudera.org:8080/13050
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>