IMPALA-8428: Bump CDH_BUILD_NUMBER to 1055188.
Brings in HADOOP-14747 (S3AInputStream to implement CanUnbuffer) which
is necessary for IMPALA-8428. The file handle cache calls `unbuffer()`
on a file handle before returning it to the cache. The call to
`unbuffer()` releases any resources that the file handle is holding onto
(in the S3A case the underlying S3ObjectInputStream is closed). Without
this fix an impalad would quickly crash as every cached file handle
would be holding onto a HTTP connection.

Testing:
* Ran an exhaustive build

Change-Id: I44b2169afff9af41d6f31312feb74ee362ecacf5
Reviewed-on: http://gerrit.cloudera.org:8080/13212
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>