IMPALA-8369 : Fix for tests failing with incompatible column changes
In Hive-3 the configuration for allowing users to make incompatible
column type changes was disabled by default. In Hive-2 this was allowed.
Some of the tests like data_errors/test_data_errors.py and
metadata/test_compute_stats.py make changes to column types which are
disallowed by HMS-3 by default. This change adds a configuration option
in hive-site.xml to allow making incompatible changes to column types so
that we can run the existing tests with HMS-3.

Also, in HMS-3 there are certain new event types (OPEN_TXN, COMMIT_TXN,
etc) which may not have dbname set. This breaks the assumption in the
code in EventProcessor which expects dbName_ to be not null at all
times. This patch also makes changes in the EventProcessor so that such
Ignored events do not fail precondition checks during event processing.

Change-Id: I488121f21d9b35d33dd003b2670bc0bbe1fee4b6
Reviewed-on: http://gerrit.cloudera.org:8080/13254
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>