IMPALA-8476: Replace Sentry admin check workaround with proper Sentry API
Impala uses a workaround to detect if a user is a Sentry admin by calling
the Sentry API to list privileges associated with the user since previously
Sentry did not provide a suitable API to peform this check.
This patch invokes a new API in SENTRY-2440 to perform the Sentry admin check.

Also modified test_sentry.py to exercise the code paths corresponding to the
following 3 different types of users: (i) a Sentry admin, (ii) an existing
user which is not a Sentry admin, and (iii) a non-existing user.

Testing:
1. Passed the tests in the revised test_sentry.py.

Change-Id: I5a27140d401494bc372ad0da96ada57bda94cd82
Reviewed-on: http://gerrit.cloudera.org:8080/13346
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>