IMPALA-8545: Test Ldap authentication
Currently, Impala does not have any automated tests for LDAP auth
functionality, due to the challenge of setting up an LDAP server
for use by the minicluster.

This patch adds LDAP tests by utilizing the ApacheDS project's
unit testing functionality, which works with JUnit to setup up
a local LDAP server for the duration of a test suite.

This requires running an Impala cluster with custom arguments to
set up LDAP auth. This patch introduces a concept of FE custom cluster
tests which must be in the package org.apache.impala.customcluster.
These tests are filtered out from the other FE tests in
bin/run-all-tests.sh and run with the other custom cluster tests so
that they don't affect other tests that expect Impala to have been
started with particular flags.

Testing:
- Ran a full core run and confirmed that new tests run as expected.

Change-Id: I92b5e60860c60209c1bd8afe5b3ea201fb7a7513
Reviewed-on: http://gerrit.cloudera.org:8080/13337
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>