IMPALA-8369 : Skip test_owner_privileges test when running against Hive-3
Currently, when running with USE_CDP_HIVE=true, Sentry service's sync
with HMS is very slow. This is most likely due to the fact that in HMS-3
the notification events are generated using the JSONMessageFactory
provided by Metastore, unlike in case of HMS-2 setup. When running
against HMS-2, Sentry provides its own MessageFactory implementation
which has its limitations and cannot be used in HMS-3. In order to fix
this Sentry should add support for the out-of-box message factory
available in Hive-3 (See SENTRY-2518).

Due to these additional delays from Sentry test_owner_privileges fails
due to race conditions between the cached information in catalog and
Sentry server (See IMPALA-8550). This patch disables this
test when running against HMS-3 until we fix the issues both on
the Sentry and Impala side.

Testing done:
1. Confirmed the test is skipped when using USE_CDP_HIVE=true
2. Confirmed the test is not skipped when using USE_CDP_HIVE=false

Change-Id: I9f904446f50b5095443bf27b3092a2e3665b76d3
Reviewed-on: http://gerrit.cloudera.org:8080/13339
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>