IMPALA-4271: Introduce the new Kudu storage handler
This commit introduces the new Kudu storage handler
'org.apache.kudu.hive.KuduStorageHandler' that is used in Kudu/HMS
integration. Tables with either the legacy storage handler
'com.cloudera.kudu.hive.KuduStorageHandler' or the new one are now
considered as Kudu tables. The actual work of adopting the new storage
handler will be addressed in a series of follow-up patches, which rely
on it to support Kudu/HMS integration in Impala.

Change-Id: I75bcd5246005f4e35251aef9219f4d07eeb87dc6
Reviewed-on: http://gerrit.cloudera.org:8080/13358
Reviewed-by: Thomas Marshall <tmarshall@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>