IMPALA-8562: Data cache should skip insertion of uncacheable ScanRanges
As shown in IMPALA-8561, there are some paths in the code which
create uncacheable ScanRanges. These uncacheable ScanRanges have
mtime of -1. 'mtime' is used for differentiating versions of files
with the same names. An mtime == -1 means the cache entry could
potentially be from any versions of a file with the same name.

This change skips lookup or insertion of ScanRange with negative
mtime, file offset or buffer length.

Testing done: Added targeted test cases in data-cache-test

Change-Id: I2294833b075a2ddcae956d9fdb04f3e85adb0391
Reviewed-on: http://gerrit.cloudera.org:8080/13369
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>