Fix bin/set-classpath.sh error handling when sourced
559f19a5be95d686a71e616517ad4ae916c85824 added an error path to
bin/set-classpath.sh which invoked 'exit'. Since this script is meant to
be sourced, rather than executed, this exit would cause the user's shell
to exit rather than just exiting the script.

This replaces the 'exit' with 'return', which has the desired behavior.
Additionally, it adds a check to make sure that it is being invoked in
the preferred manner.

Change-Id: Ic8129444df3e6907a64c40409ebc5197dca47801
Reviewed-on: http://gerrit.cloudera.org:8080/13391
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>