IMPALA-8542. Add an access trace for the data cache
This adds a relatively simple JSON-formatted access trace for the data
cache feature. Each partition stores a trace file named 'trace.txt',
with each line representing a hit, miss, or store into the cache.

The trace is collected using the kudu::AsyncLogger class which handles
buffering and deferring the actual IO to a background thread.

By default, the full cache key info is written to the trace (including
the file paths), but a flag can enable anonymization (128-bit
city-hashing) of the file names in case any user would like to capture a
trace to be shared publically without divulging their table names.

Change-Id: I2302c19abb5db19f1d3d1cd727a82977a9e2ba9c
Reviewed-on: http://gerrit.cloudera.org:8080/13425
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Michael Ho <kwho@cloudera.com>