IMPALA-8588: Fix revoke grant option with Ranger
Previously, the REVOKE GRANT OPTION type statements would not only
revoke the grant option, but also the privilege as well. The behavior
has been updated to only revoke the grant option. In Ranger UI, this is
seen as the delegate admin option. Examples:

REVOKE SELECT ON DATABASE <database> FROM USER <user>

This will revoke the SELECT privilege on the database resource, but
if there are other privileges on that resource the grant option will
remain for those privileges.

REVOKE GRANT OPTION FOR SELECT ON DATABASE <database> FROM USER <user>

This will revoke the grant option for all privileges on this database
resource. It will not revoke the SELECT privilege on the resource.

Testing:
- Ran all FE tests
- Ran all E2E tests
- Updated test_ranger to test behavior for when REVOKE GRANT OPTION
  statements are submitted.

Change-Id: Iddfccb442c3be3c266dbc2d8ae85c5674c534d7c
Reviewed-on: http://gerrit.cloudera.org:8080/13450
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>