IMPALA-8505: Disallow setting Kudu table id in ALTER TABLE
This commit disallows explicitly setting the Kudu table id property for
Kudu tables in a ALTER TABLE set statement. The Kudu table id property
is generated when the table is created and should not be altered
afterwards.

This commit also extracts Kudu-related analyzing tests for table alteration,
so that they can be run with or without Kudu/HMS integration enabled.

Change-Id: Ibc1f3f87054db5e28f72a5b54ea5b3c040c1bf22
Reviewed-on: http://gerrit.cloudera.org:8080/13465
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Thomas Marshall <tmarshall@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>