IMPALA-8459. LocalCatalog: Allow dropping tables that fail to load
This adds support in LocalCatalog for IncompleteTables. A new
FeIncompleteTable interface is introduced, and when a table fails to
load, it produces a FeIncompleteTable instead of an immediate error.

This allows DROP TABLE statements to get past analysis and execute even
when tables are in bad states (eg a Kudu table with no backing table in
Kudu itself).

This re-enables some tests that were previously disabled for
LocalCatalog.

Change-Id: Ia85d6b10669866dcc9f2bf7385a4775e483dd6b0
Reviewed-on: http://gerrit.cloudera.org:8080/13557
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>