IMPALA-8697: fix assert in test_query_options for EC test runs
IMPALA-7290 added a timezone query option check to this test but the
expected order of query options being printed in the query profile was
wrong specifically for an option expected in a test run with erasure
coding turned on. This patch just puts them in the right order in
the expected string.

Testing:
Tested on EC build.

Change-Id: I50a3fffa8c4800cc159bf512692393c6b4392cf9
Reviewed-on: http://gerrit.cloudera.org:8080/13709
Reviewed-by: Bikramjeet Vig <bikramjeet.vig@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>