IMPALA-8700: Use int64_t mtime for HdfsScanNodeBase::AllocateScanRange()
HDFS defines the modification time as the milliseconds since Jan 1st,
1970. Since milliseconds wrap every 49 days in a 4 byte integer,
all locations storing modification time need to use an 8 byte integer.
The overload of HdfsScanNodeBase::AllocateScanRange() that is called
from Impala-lzo uses a 4 byte integer. When this wraps, it triggers
a DCHECK in ScanRange::Reset() that assumes mtime > 0.

This changes the HdfsScanNodeBase::AllocateScanRange() overload to
take an int64_t. The Impala-lzo side already supplies an int64_t.

Change-Id: Icce5b22361c74333f3213dce629e0e625550573f
Reviewed-on: http://gerrit.cloudera.org:8080/13712
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>